requirements.txt
asgiref==3.9.1
certifi==2025.10.5
channels==4.3.1
charset-normalizer==3.4.4
decorator==5.2.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.3.3
pillow==11.3.0
proglog==0.1.12
pymongo==3.11.4
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
tqdm==4.67.1
urllib3==2.5.0
